+++ lib/image_util.rb
@@ -15,5 +15,6 @@ module ImageUtil
   autoload :Statistic, "image_util/statistic"
   autoload :Terminal, "image_util/terminal"
   autoload :View, "image_util/view"
+  autoload :CLI, "image_util/cli"
   autoload :Benchmarking, "image_util/benchmarking"
 end